02-nosql-drinks-api/requirements.txt
asgiref==3.8.1
boto3==1.37.13
botocore==1.37.13
distlib==0.3.9
Django==5.1.7
djangorestframework==3.15.2
filelock==3.17.0
jmespath==1.0.1
platformdirs==4.3.6
python-dateutil==2.9.0.post0
s3transfer==0.11.4
six==1.17.0
sqlparse==0.5.3
tzdata==2025.1
urllib3==2.3.0
virtualenv==20.29.3
